add-dhcp-server-editor-v1

+++ Form1.cs
@@ -27,7 +27,7 @@ namespace MikrotikAddClient
             DHCP_Server = DHCP_Value.Text;
 
             clientInfoPrint.Text = "/queue simple\n" +
-                "add burst-limit=" + ((int)(UploadSpeed * 1.1)) + "M/" + DownloadSpeed * 1.1 + "M burst-threshold=" + UploadSpeed + "M/" + DownloadSpeed + "M burst-time=8s/8s max-limit=" + UploadSpeed + "M/" + DownloadSpeed + "M name=" + clientDescription + " queue=wireless-default/wireless-default target=" + ipAddress +
+                "add burst-limit=" + ((int)(UploadSpeed * 1.1)) + "M/" + ((int)(DownloadSpeed * 1.1)) + "M burst-threshold=" + UploadSpeed + "M/" + DownloadSpeed + "M burst-time=8s/8s max-limit=" + UploadSpeed + "M/" + DownloadSpeed + "M name=" + clientDescription + " queue=wireless-default/wireless-default target=" + ipAddress +
                 "\n/ip firewall address-list\nadd address=" + ipAddress + " comment=" + clientDescription + " list=klienci" +
                 "\n/ip dhcp-server lease\nadd address=" + ipAddress + " always-broadcast=yes comment=" + clientDescription + " disabled=no mac-address=" + MAC_Address + " server=" + DHCP_Server;
         }
